Authorization management

+++ nova/src/Authorize/Http/Controllers/PermissionsController.php
@@ -1,5 +1,6 @@
 <?php namespace Nova\Authorize\Http\Controllers;
 
+use Nova\Authorize\Role;
 use Illuminate\Http\Request;
 use Nova\Authorize\Permission;
 use Nova\Foundation\Http\Controllers\Controller;
@@ -15,23 +16,26 @@ class PermissionsController extends Controller
 
 	public function index()
 	{
-		// Authorize
+		$roleClass = new Role;
+		$permissionClass = new Permission;
+
+		$this->authorize('manage', $permissionClass);
 
 		$permissions = Permission::get();
 
-		return view('pages.authorize.all-permissions', compact('permissions'));
+		return view('pages.authorize.all-permissions', compact('permissions', 'roleClass', 'permissionClass'));
 	}
 
 	public function create()
 	{
-		// Authorize
+		$this->authorize('create', new Permission);
 
 		return view('pages.authorize.create-permission');
 	}
 
 	public function store(Request $request)
 	{
-		// Authorize
+		$this->authorize('create', new Permission);
 
 		$this->validate($request, [
 			'name' => 'required',
@@ -45,14 +49,14 @@ class PermissionsController extends Controller
 
 	public function edit(Permission $permission)
 	{
-		// Authorize
+		$this->authorize('update', $permission);
 
 		return view('pages.authorize.update-permission', compact('permission'));
 	}
 
 	public function update(Request $request, Permission $permission)
 	{
-		// Authorize
+		$this->authorize('update', $permission);
 
 		$this->validate($request, [
 			'name' => 'required',
@@ -66,8 +70,10 @@ class PermissionsController extends Controller
 
 	public function destroy(Permission $permission)
 	{
-		// Authorize
+		$this->authorize('delete', $permission);
 
+		$permission->roles()->sync([]);
+		
 		$permission->delete();
 
 		return redirect()->route('permissions.index');
